+++ zebra-chain/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `tests::vectors::test_block_sapling_roots`
+
 ## [3.0.0] - 2025-10-15
 
 In this release we removed a significant amount of Sapling-related code in favor of upstream implementations.